fix(migration): use DOUBLE PRECISION instead of DOUBLE in V37 (PostgreSQL)

+++ backend/src/main/resources/db/migration/V37__add_dual_distance_to_rep_trips.sql
@@ -5,6 +5,6 @@
 --   distance_source = which value populated total_km ("actual",
 --                     "ideal_fallback", "haversine", "manual", ...)
 ALTER TABLE rep_trip
-    ADD COLUMN ideal_km        DOUBLE NULL,
-    ADD COLUMN actual_km       DOUBLE NULL,
+    ADD COLUMN ideal_km        DOUBLE PRECISION NULL,
+    ADD COLUMN actual_km       DOUBLE PRECISION NULL,
     ADD COLUMN distance_source VARCHAR(32) NULL;
